requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.4
cffi==1.11.5
cryptography==2.3
Django==2.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-tables2==2.0.0a5
django-yarn==0.0.1
idna==2.7
kombu==4.2.1
Markdown==2.6.11
mysqlclient==1.3.13
pycparser==2.18
PyMySQL==0.9.2
pytz==2018.5
six==1.11.0
vine==1.1.4